+++ truthbeam/internal/metadata/metadata.go
@@ -1,9 +1,0 @@
-package metadata
-
-import "go.opentelemetry.io/collector/component"
-
-var Type = component.MustNewType("truthbeam")
-
-const (
-	LogsStability = component.StabilityLevelAlpha
-)